+++ web/src/components/layout/AppMenu.tsx
@@ -2,10 +2,10 @@ import AddCircleIcon from "@mui/icons-material/AddCircle";
 import Brightness4Icon from "@mui/icons-material/Brightness4";
 import Brightness7Icon from "@mui/icons-material/Brightness7";
 import CheckIcon from "@mui/icons-material/Check";
-import MenuIcon from "@mui/icons-material/Menu";
-import TuneIcon from "@mui/icons-material/Tune";
 import LoginIcon from "@mui/icons-material/Login";
 import LogoutIcon from "@mui/icons-material/Logout";
+import MenuIcon from "@mui/icons-material/Menu";
+import TuneIcon from "@mui/icons-material/Tune";
 import {
   Box,
   Divider,
@@ -21,11 +21,11 @@ import {
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { generatePath, useLocation, useNavigate } from "react-router-dom";
+import { JiraInfoDialog } from "../../jira/components/JiraInfoDialog";
+import { useIsJiraAuthenticated } from "../../jira/jira-api";
+import { keijoJiraApiUrl } from "../../jira/jiraConfig";
 import useDarkMode from "../../theme/useDarkMode";
 import LabelledIconButton from "../LabelledIconButton";
-import { keijoJiraApiUrl } from "../../jira/jiraConfig";
-import { useIsJiraAuthenticated } from "../../jira/jiraApi";
-import { JiraInfoDialog } from "../../jira/components/JiraInfoDialog";
 
 const AppMenuButton = () => {
   const navigate = useNavigate();
